Moved onboarding page 2 to it's on component and created a story for it and also onboardinglayout2
