+++ packages/web/stories/OnboardingPage2Story.stories.tsx
@@ -1,0 +1,13 @@
+import { ComponentStory, ComponentMeta } from '@storybook/react'
+import OnboardingPage2 from '../components/templates/onboarding/02'
+
+export default {
+  title: 'Components/OnboardingPage2Story',
+  component: OnboardingPage2,
+} as ComponentMeta<typeof OnboardingPage2>
+
+export const OnboardingPage2Story: ComponentStory<typeof OnboardingPage2> = (args: any) => {
+  return (
+    <OnboardingPage2 />
+  )
+}
